Add per-minute support, compartmentalize
