+++ packages/musicxml-parser/src/parser.ts
@@ -1,39 +1,4 @@
-import JSZip from "jszip";
-
-/**
- * Represents a beat in a musical performance or composition.
- * Provides details about the beat's position, duration, and optional notes.
- *
- * Includes only specific properties needed for parsing functionality.
- */
-export interface Beat {
-    /** Duration from this beat to the next in seconds. This is derived from tempo */
-    duration: number;
-    /** Human-readable notes about this beat. These are not the musical "notes" that are played. */
-    notes?: string;
-}
-
-/**
- * Represents a musical measure in a musical composition.
- * Contains information about the structure and content of a specific measure.
- *
- * Includes only specific properties needed for parsing functionality.
- */
-export interface Measure {
-    /** The measure's number in the piece. Unique and integer */
-    number: number;
-    /** Optional rehearsal mark for the measure. I.e. "Big Box A" or "Measure 128" */
-    rehearsalMark?: string;
-    /** Human-readable notes about the measure. These are not the musical "notes" that are played. */
-    notes?: string;
-    /** The beats that belong to this measure */
-    beats: Beat[];
-}
-
-// 60 seconds in a minute, divided by tempo in beats per minute
-export function secondsPerQuarterNote(tempo: number): number {
-    return 60 / tempo;
-}
+import { type Measure, type Beat, secondsPerQuarterNote } from "./utils.ts";
 
 /**
  * Parses a MusicXML string and converts it into an array of musical measures.
@@ -43,7 +8,7 @@ export function secondsPerQuarterNote(tempo: number): number {
 export function parseMusicXml(xmlText: string): Measure[] {
     const measures: Measure[] = [];
     let pos = 0;
-    let tempo: number = 0;
+    let tempo: number = 120;
     let timeSignature: string = "0/0";
 
     /** Given a time signature, bigBeats maps it to the:
@@ -64,7 +29,7 @@ export function parseMusicXml(xmlText: string): Measure[] {
         "6/8": [2, 1.5],
         "9/8": [3, 1.5],
         "12/8": [4, 1.5],
-        "7/8": [7, 1], // Likely this will end up being 3 "big beats" (for situations like 2+2+3)
+        "7/8": [7, 1],
     };
 
     // If the XML has multiple parts, stop after the first part
@@ -97,6 +62,13 @@ export function parseMusicXml(xmlText: string): Measure[] {
         if (tempoMatches.length > 0) {
             const lastTempo = tempoMatches[tempoMatches.length - 1];
             tempo = parseInt(lastTempo![1] as string);
+        } else {
+            const tempoMatch = measureText.match(
+                /<per-minute[^>]*>(.*?)<\/per-minute>/s,
+            );
+            if (tempoMatch && tempoMatch[1]) {
+                tempo = parseInt(tempoMatch[1]);
+            }
         }
 
         // Update time signature if new one exists
@@ -122,7 +94,7 @@ export function parseMusicXml(xmlText: string): Measure[] {
         // Check for valid time signature
         const [bigBeatCount, quarterNotesPerBigBeat] = bigBeats[
             timeSignature
-        ] ?? [0, 0];
+        ] ?? [1, 1];
         if (bigBeatCount === 0) {
             throw new Error(`Unsupported time signature: ${timeSignature}`);
         }
@@ -152,67 +124,3 @@ export function parseMusicXml(xmlText: string): Measure[] {
 
     return measures;
 }
-
-/**
- * Extracts the XML content (as a string) from a MusicXML (.mxl) file.
- * This function reads the .mxl file, unzips it, and returns the content of the first XML file found.
- * Assumes that the zipped .mxl file contains only one desired XML file.
- */
-export async function extractXmlFromMxlFile(
-    fileBuffer: ArrayBuffer,
-): Promise<string> {
-    // load file
-    const zip = await JSZip.loadAsync(fileBuffer);
-
-    // loop through files in zip
-    for (const fileName of Object.keys(zip.files)) {
-        // locate a .xml or .musicxml file that is not container.xml or in META-INF
-        if (
-            (fileName.endsWith(".xml") || fileName.endsWith(".musicxml")) &&
-            !fileName.endsWith("container.xml") &&
-            !fileName.startsWith("META-INF/")
-        ) {
-            return await zip.files[fileName]!.async("text");
-        }
-    }
-
-    // unzipped, but no xml or musicxml file found
-    throw new Error("No XML file found.");
-}
-
-/**
- * Parses score xml as a string from an MXL file
- * @param filePath path to .mxl file on disk
- * @returns parsed XML or undefined if an error occurred
- *
- * Legacy function to parse MXL files.
- */
-/*
-export function parseMxl(filePath: string): string | undefined {
-    try {
-        var zip = new AdmZip(filePath);
-        var zipEntries = zip.getEntries();
-
-        // Find the root container, this will tell us which zip entry contains the actual score data we care about.
-        const rootContainer = zipEntries.find(
-            (entry: any) => entry.entryName === "META-INF/container.xml",
-        );
-
-        // Parse the root container data to a string
-        const rootContainerData = rootContainer.getData().toString("utf8");
-
-        // Pull out the root file path using a regex (I gave up on parsing this "correctly" using xml)
-        const regex = /rootfile full-path="([^"]*)"/;
-        const match = rootContainerData.match(regex);
-        let scorePath = match ? match[1] : undefined;
-
-        // Find the score container and create a string from its xml data
-        let scoreContainer = zipEntries.find(
-            (zipEntry: any) => zipEntry.entryName === scorePath,
-        );
-        return scoreContainer.getData().toString("utf8");
-    } catch (error) {
-        console.error("Error parsing MXL:", error);
-        return undefined;
-    }
-}*/
